modified paymentGateway adding phonePe

+++ src/order/v2/cart/v2/cart.controller.js
@@ -1,4 +1,6 @@
 import CartService from './cart.service.js';
+import {validationResult} from "express-validator"
+import BadRequestParameterError from '../../../../lib/errors/bad-request-parameter.error.js';
 
 const cartService = new CartService();
 
@@ -7,8 +9,11 @@ class CartController {
     async addItem(req, res, next) {
         try {
 
-           return  res.send(await cartService.addItem({...req.body,...req.params}));
-
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                throw new BadRequestParameterError(errors.array()[0].msg);
+            }
+           return res.send(await cartService.addItem({...req.body, userId: req.user.decodedToken.uid}));
         }
         catch (err) {
            next(err);
@@ -17,30 +22,23 @@ class CartController {
 
     async getCartItem(req, res, next) {
         try {
-
-            return  res.send(await cartService.getCartItem({...req.body,...req.params}));
-
+            return  res.send(await cartService.getCartItem({...req.body,userId: req.user.decodedToken.uid}));
         }
         catch (err) {
             next(err);
         }
     }
 
-
-    async getAllCartItem(req, res, next) {
-        try {
-
-            return  res.send(await cartService.getAllCartItem({...req.body,...req.params}));
-
-        }
-        catch (err) {
-            next(err);
-        }
-    }
 
     async updateItem(req, res, next) {
         try {
-            return  res.send(await cartService.updateItem({...req.body,...req.params}));
+
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                throw new BadRequestParameterError(errors.array()[0].msg);
+            }
+
+            return  res.send(await cartService.updateItem({...req.body, cartItemId: req?.params?.itemId, userId: req.user.decodedToken.uid }));
         }
         catch (err) {
             next(err);
@@ -49,7 +47,11 @@ class CartController {
 
     async removeItem(req, res, next) {
         try {
-            return  res.send(await cartService.removeItem({...req.body,...req.params}));
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                throw new BadRequestParameterError(errors.array()[0].msg);
+            }
+            return res.send(await cartService.removeItem({...req.body,...req.params}));
 
         }
         catch (err) {
@@ -59,7 +61,20 @@ class CartController {
 
     async clearCart(req, res, next) {
         try {
-            return  res.send(await cartService.clearCart({...req.body,...req.params}));
+            const errors = validationResult(req);
+            if (!errors.isEmpty()) {
+                throw new BadRequestParameterError(errors.array()[0].msg);
+            }
+            return  res.send(await cartService.clearCart({...req.body,...req.params , userId: req.user.decodedToken.uid }));
+        }
+        catch (err) {
+            next(err);
+        }
+    }
+
+    async getAllCartItem(req, res, next) {
+        try {
+            return  res.send(await cartService.getAllCartItem({...req.body,...req.params}));
         }
         catch (err) {
             next(err);
